+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Bugfixes:
 - Fixed gauges may not save ingredient amount change if source gauge relocated.
 - Fixed gauge timer animation is not properly rendered with some animation lib mods installed.
 - Fixed controller GUI may not open in forge-1.20 if data is too large.
-- Fixed redstone link component does not function properly when connect to another sable sub-level.
+- Fixed some redstone link control bugs.
 
 ## 1.1.1